Fix compute precision

+++ gemma_chat/export.py
@@ -235,7 +235,10 @@ def _hlo_to_mlpackage(
                 mil_program,
                 source="milinternal",
                 minimum_deployment_target=ct.target.iOS18,
-                compute_precision=ct.precision.FLOAT16,
+                # When using ct.precision.FLOAT16 the model output becomes unstable and garbage
+                # tokens are produced. We use FLOAT16 precision in the model where permissible
+                # by manual casting.
+                compute_precision=ct.precision.FLOAT32,
                 pass_pipeline=pipeline,
                 skip_model_load=True,
             )
